Create benchmarkpseudocode4SweepingSubnet
We have now established a quantitative bound on the radius~$r$ in terms of the resolution~$\xi$. For simplicity, we restrict the scope of our following theorems to configurations where this radius exactly matches the radius~$r_{\text{max}}$ of a sphere~$\mathcal{S}_r$ that is tangent to the light cone. In this case, the sweeping subnet of the causal barrier can immediately be converted into a sweeping subnet of~$\mathcal{S}_r$ by restricting both manifolds to their intersection. The union of these sweeping subnets indeed corresponds to an optimal tessellation for tracing the unique maximal rays that leave the angular position~$\vec{x}$.